(PC-15946)[API] chore: mutualize regions name retrieval
